_requirements.txt
biopython==1.80
dash==2.14.1
gdown==4.7.1
gwosc==0.7.1
gwpy==3.0.7
matplotlib==3.6.2
numpy==1.24.0
pandas==1.5.3
panel==0.14.2
param==1.12.3
Pillow==9.4.0
Pillow==10.1.0
plotly==5.11.0
plotly_express==0.4.1
psutil==5.9.4
pyvista==0.38.2
Requests==2.31.0
scikit_learn==1.2.2
scipy==1.9.3
seaborn==0.13.0
setuptools==67.6.0
tqdm==4.64.1
wheel==0.38.4
